add prototype of rpc like

+++ Socket/Serializer.cpp
@@ -26,7 +26,7 @@ std::ostream& operator<<(std::ostream& output,const Serializer& self)
 {
     output<<"["<<self.size()<<"]";
     for(int i=self._cursor_begin; i<self._cursor_end;++i)
-        output<<"<"<<self._buffer[i]<<">";
+        output<<"<"<<(int)self._buffer[i]<<">";
     return output;
 };
